revert

+++ src/main/java/com/watchworthy/api/dto/AwardDTO.java
@@ -9,7 +9,9 @@ public class AwardDTO {
     private String description;
     private Long movieId;
 
+    // Constructors, getters, and setters
 
+    // Constructors
     public AwardDTO() {
     }
 
@@ -72,7 +74,7 @@ public class AwardDTO {
         this.description = description;
     }
 
-    public Integer getMovieId() {
+    public Long getMovieId() {
         return movieId;
     }
